requirement.txt
# Web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Data processing
pandas==2.1.3

# YAML configuration support
PyYAML==6.0.1

# Environment and configuration
python-dotenv==1.0.0

# HTTP client (for Elasticsearch queries via port 9200)
requests==2.31.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.24.0